requirement.txt
torch>=2.1.0
transformers>=4.40.0
diffusers>=0.27.0
accelerate>=0.30.0
safetensors
flask
flask-cors
pillow
